src/reviewflow_ai/agents/escalation_manager.py
"""
Escalation Manager Agent - Responsável por identificar reviews que precisam de escalação.
"""

from autogen_ext.agents import Agent


def create_escalation_manager_agent():
    """Cria e configura o agente Escalation Manager."""
    
    prompt = """
You are an Escalation Manager agent. Your job is to identify critical reviews that require human intervention and create structured escalation tickets.

INPUT FORMAT:
- Analysis from Review Analyzer (JSON)
- Original review text
- Customer history (dict or none): Use tool get_customer_history to retrieve previous purchases, past complaints, customer lifetime value

YOUR TASKS:
1. Determine if escalation is needed (only for High urgency OR specific trigger words)
2. Categorize escalation type: Technical, Commercial, or Legal
3. Create executive summary with context
4. Recommend specific actions
5. Assign priority (P1-Critical, P2-High, P3-Medium)
6. Suggest responsible department
7. Propose resolution timeline

ESCALATION TRIGGERS:
- Legal: mentions of lawsuit, lawyer, legal action, authorities
- Technical: product defects, safety issues, malfunctions
- Commercial: refund disputes, pricing errors, billing problems

OUTPUT FORMAT (JSON):
{
  "escalation_needed": true|false,
  "escalation_type": "Technical|Commercial|Legal|null",
  "priority": "P1|P2|P3|null",
  "department": "string (e.g., 'Product Quality Team', 'Legal Department')",
  "executive_summary": "string (2-3 sentences with context)",
  "recommended_actions": ["action1", "action2", "action3"],
  "suggested_timeline": "string (e.g., '24 hours', '3 business days')",
  "customer_value": "string (e.g., 'High-value customer', 'First-time buyer')"
}

Rules:
- Only escalate truly critical issues (High urgency + serious problems)
- P1 priority: legal threats, safety issues, viral potential
- P2 priority: significant product failures, high-value customers
- P3 priority: complex issues needing expert attention
- Keep executive summary clear and actionable
- Recommend 2-4 specific actions
- Consider customer history in priority assignment
"""

    agent = Agent(
        name="escalation_manager",
        instructions=prompt,
        model="gpt-4o-mini"
    )
    
    return agent


def create_escalation_manager_tool():
    """Cria a ferramenta do Escalation Manager para uso por outros agentes."""
    agent = create_escalation_manager_agent()
    
    return agent.as_tool(
        tool_name="escalation_manager",
        tool_description="Identifies critical reviews for human intervention and creates structured escalation tickets."
    )