requirements-production.txt
# Dependências essenciais para ReviewFlow AI
fastapi==0.110.0
uvicorn[standard]==0.27.0
pydantic==2.6.1
python-dotenv==1.0.1
httpx==0.26.0

# OpenAI para integração direta
openai==1.12.0

# Dependências básicas
aiofiles==0.8.0
nest-asyncio==1.6.0

# Para desenvolvimento e teste
pytest==8.0.0
pytest-asyncio==0.23.4